Simplify post-rebase fixes
