+++ crates/uv-extract/src/sync.rs
@@ -173,3 +173,28 @@ pub fn unzip(reader: fs_err::File, target: &Path) -> Result<Vec<(PathBuf, u64)>,
         .filter_map(Result::transpose)
         .collect::<Result<_, Error>>()
 }
+
+/// Extract the top-level directory from an unpacked archive.
+///
+/// The specification says:
+/// > A .tar.gz source distribution (sdist) contains a single top-level directory called
+/// > `{name}-{version}` (e.g. foo-1.0), containing the source files of the package.
+///
+/// This function returns the path to that top-level directory.
+pub fn strip_component(source: impl AsRef<Path>) -> Result<PathBuf, Error> {
+    // TODO(konstin): Verify the name of the directory.
+    let top_level = fs_err::read_dir(source.as_ref())
+        .map_err(Error::Io)?
+        .collect::<std::io::Result<Vec<fs_err::DirEntry>>>()
+        .map_err(Error::Io)?;
+    match top_level.as_slice() {
+        [root] => Ok(root.path()),
+        [] => Err(Error::EmptyArchive),
+        _ => Err(Error::NonSingularArchive(
+            top_level
+                .into_iter()
+                .map(|entry| entry.file_name())
+                .collect(),
+        )),
+    }
+}
